+++ app/src/main/java/com/inspur/emmcloud/ui/schedule/task/TaskListFragment.java
@@ -2,14 +2,13 @@ package com.inspur.emmcloud.ui.schedule.task;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v4.widget.SwipeRefreshLayout;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
+import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.ListView;
 import android.widget.TextView;
@@ -18,45 +17,35 @@ import com.inspur.emmcloud.R;
 import com.inspur.emmcloud.adapter.TaskListAdapter;
 import com.inspur.emmcloud.api.APIInterfaceInstance;
 import com.inspur.emmcloud.api.apiservice.ScheduleApiService;
+import com.inspur.emmcloud.bean.schedule.task.GetTaskListResult;
+import com.inspur.emmcloud.bean.schedule.task.Task;
 import com.inspur.emmcloud.bean.system.SimpleEventMessage;
-import com.inspur.emmcloud.bean.work.GetTaskListResult;
-import com.inspur.emmcloud.bean.work.Task;
 import com.inspur.emmcloud.config.Constant;
 import com.inspur.emmcloud.util.common.NetUtils;
 import com.inspur.emmcloud.util.common.PreferencesUtils;
 import com.inspur.emmcloud.util.common.StringUtils;
 import com.inspur.emmcloud.util.privates.WebServiceMiddleUtils;
 import com.inspur.emmcloud.widget.MySwipeRefreshLayout;
-import com.inspur.emmcloud.widget.dialogs.MyQMUIDialog;
-import com.qmuiteam.qmui.widget.dialog.QMUIDialog;
-import com.qmuiteam.qmui.widget.dialog.QMUIDialogAction;
+import com.inspur.emmcloud.widget.dialogs.CustomDialog;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 import org.greenrobot.eventbus.ThreadMode;
-import org.xutils.view.annotation.ContentView;
-import org.xutils.view.annotation.ViewInject;
-import org.xutils.x;
 
 import java.util.ArrayList;
 
 /**
  * Created by yufuchang on 2019/4/1.
  */
-@ContentView(R.layout.fragment_task_list)
 public class TaskListFragment extends Fragment {
 
     public static final String TASK_TASK_ENTITY = "task";
     public static final String TASK_CURRENT_INDEX = "tabIndex";
-    @ViewInject(R.id.lv_task)
     private ListView taskListView;
-    @ViewInject(R.id.refresh_layout)
     private MySwipeRefreshLayout swipeRefreshLayout;
-    @ViewInject(R.id.ll_no_search_result)
     private LinearLayout noSearchResultLayout;
-    @ViewInject(R.id.tv_no_result)
     private TextView noResultText;
-    private boolean injected = false;
+    private ImageView taskNoResultImageView;
     private String orderBy = "PRIORITY";
     private String orderType = "ASC";
     private int deletePosition = -1;
@@ -68,28 +57,29 @@ public class TaskListFragment extends Fragment {
     private ArrayList<Task> taskList = new ArrayList<Task>();
     private boolean isPullUp = false;
     private int page = 0;
+    private View rootView;
 
 
     @Override
-    public void onCreate(@Nullable Bundle savedInstanceState) {
+    public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
+        rootView = LayoutInflater.from(getActivity()).inflate(R.layout.fragment_task_list, null);
+        initViews();
         EventBus.getDefault().register(this);
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        injected = true;
-        return x.view().inject(this, inflater, container);
-    }
-
-    @Override
-    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
-        super.onViewCreated(view, savedInstanceState);
-        if (!injected) {
-            x.view().inject(this, this.getView());
+        if (rootView == null) {
+            rootView = inflater
+                    .inflate(R.layout.fragment_task_list, container, false);
         }
-        initViews();
+        ViewGroup parent = (ViewGroup) rootView.getParent();
+        if (parent != null) {
+            parent.removeView(rootView);
+        }
+        return rootView;
     }
 
     @Subscribe(threadMode = ThreadMode.MAIN)
@@ -109,23 +99,27 @@ public class TaskListFragment extends Fragment {
      * @param searchContent
      */
     public void setSearchContent(String searchContent) {
-        swipeRefreshLayout.setCanLoadMore(StringUtils.isBlank(searchContent));
-        swipeRefreshLayout.setEnabled(StringUtils.isBlank(searchContent));
+        if (swipeRefreshLayout != null) {
+            swipeRefreshLayout.setCanLoadMore(StringUtils.isBlank(searchContent));
+            swipeRefreshLayout.setEnabled(StringUtils.isBlank(searchContent));
+        }
         this.searchContent = searchContent;
         if (adapter != null) {
             searchTaskListBySearchContent();
         }
     }
 
-    /**
-     * 告知Fragment当前索引
-     *
-     * @param currentIndex
-     */
-    public void setCurrentIndex(int currentIndex) {
-        this.currentIndex = currentIndex;
-        swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE);
-    }
+//    /**
+//     * 告知Fragment当前索引
+//     *
+//     * @param currentIndex
+//     */
+//    public void setCurrentIndex(int currentIndex) {
+//        this.currentIndex = currentIndex;
+//        if(swipeRefreshLayout != null){
+//            swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE && (uiTaskList.size() % 12 == 0));
+//        }
+//    }
 
     /**
      * 根据搜索内容搜索列表
@@ -141,6 +135,12 @@ public class TaskListFragment extends Fragment {
     }
 
     private void initViews() {
+
+        taskListView = rootView.findViewById(R.id.lv_task);
+        swipeRefreshLayout = rootView.findViewById(R.id.refresh_layout);
+        noSearchResultLayout = rootView.findViewById(R.id.ll_no_search_result);
+        noResultText = rootView.findViewById(R.id.tv_no_result);
+        taskNoResultImageView = rootView.findViewById(R.id.iv_task_no_result);
         currentIndex = getArguments().getInt(TaskFragment.MY_TASK_TYPE, TaskFragment.MY_MINE);
         apiService = new ScheduleApiService(getActivity());
         apiService.setAPIInterface(new WebService());
@@ -167,24 +167,24 @@ public class TaskListFragment extends Fragment {
      */
     private void initPullRefreshLayout() {
         //已完成页面设置可以上拉加载
-        swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE);
+        swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE && (uiTaskList.size() % 12 == 0));
         swipeRefreshLayout.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
             @Override
             public void onRefresh() {
                 isPullUp = false;
                 page = 0;
-                if (NetUtils.isNetworkConnected(getActivity())) {
+                if (NetUtils.isNetworkConnected(getActivity(), false)) {
                     getCurrentTaskList();
                 } else {
                     swipeRefreshLayout.setLoading(false);
                 }
-                swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE);
+                swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE && (uiTaskList.size() % 12 == 0));
             }
         });
         swipeRefreshLayout.setOnLoadListener(new MySwipeRefreshLayout.OnLoadListener() {
             @Override
             public void onLoadMore() {
-                if (NetUtils.isNetworkConnected(getActivity())) {
+                if (NetUtils.isNetworkConnected(getActivity(), false)) {
                     apiService.getFinishTasks(page, 12, "REMOVED");
                     isPullUp = true;
                 } else {
@@ -196,7 +196,7 @@ public class TaskListFragment extends Fragment {
 
     private void getCurrentTaskList() {
         currentIndex = getArguments().getInt(TaskFragment.MY_TASK_TYPE, TaskFragment.MY_MINE);
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             if (currentIndex == TaskFragment.MY_MINE) {
                 getMineTasks();
             } else if (currentIndex == TaskFragment.MY_INVOLVED) {
@@ -215,7 +215,7 @@ public class TaskListFragment extends Fragment {
      * 获取关注的任务
      */
     protected void getFocusedTasks() {
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             swipeRefreshLayout.setRefreshing(true);
             apiService.getFocusedTasks(orderBy, orderType);
         }
@@ -225,7 +225,7 @@ public class TaskListFragment extends Fragment {
      * 获取我参与的任务
      */
     protected void getInvolvedTasks() {
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             swipeRefreshLayout.setRefreshing(true);
             apiService.getInvolvedTasks(orderBy, orderType);
         }
@@ -235,7 +235,7 @@ public class TaskListFragment extends Fragment {
      * 获取我的任务
      */
     protected void getMineTasks() {
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             swipeRefreshLayout.setRefreshing(true);
             apiService.getMineTasks(orderBy, orderType);
         }
@@ -245,7 +245,7 @@ public class TaskListFragment extends Fragment {
      * 获取所有任务
      */
     private void getAllFinishTasks() {
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             swipeRefreshLayout.setRefreshing(true);
             apiService.getFinishTasks(0, 12, "REMOVED");
         }
@@ -258,7 +258,7 @@ public class TaskListFragment extends Fragment {
      * @param position
      */
     protected void deleteTasks(int position) {
-        if (NetUtils.isNetworkConnected(getActivity())) {
+        if (NetUtils.isNetworkConnected(getActivity(), false)) {
             swipeRefreshLayout.setRefreshing(true);
             apiService.setTaskFinishById(uiTaskList.get(position).getId());
             deletePosition = position;
@@ -279,20 +279,12 @@ public class TaskListFragment extends Fragment {
         public boolean onItemLongClick(AdapterView<?> parent, View view,
                                        final int position, long id) {
             if (currentIndex == 0 || currentIndex == 1) {
-                new MyQMUIDialog.MessageDialogBuilder(getActivity())
+                new CustomDialog.MessageDialogBuilder(getActivity())
                         .setMessage(R.string.mession_set_finish)
-                        .addAction(getString(R.string.cancel), new QMUIDialogAction.ActionListener() {
-                            @Override
-                            public void onClick(QMUIDialog dialog, int index) {
-                                dialog.dismiss();
-                            }
-                        })
-                        .addAction(getString(R.string.ok), new QMUIDialogAction.ActionListener() {
-                            @Override
-                            public void onClick(QMUIDialog dialog, int index) {
-                                dialog.dismiss();
-                                deleteTasks(position);
-                            }
+                        .setNegativeButton(getString(R.string.cancel), (dialog, index) -> dialog.dismiss())
+                        .setPositiveButton(getString(R.string.ok), (dialog, index) -> {
+                            dialog.dismiss();
+                            deleteTasks(position);
                         })
                         .show();
 
@@ -327,15 +319,16 @@ public class TaskListFragment extends Fragment {
                 swipeRefreshLayout.setLoading(false);
                 page = page + 1;
                 taskList.addAll(getTaskListResult.getTaskList());
-                swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE && getTaskListResult.getTaskList().size() >= 12);
             } else {
                 swipeRefreshLayout.setRefreshing(false);
                 taskList = getTaskListResult.getTaskList();
             }
-            noResultText.setVisibility(taskList.size() > 0 ? View.GONE : View.VISIBLE);
             uiTaskList.clear();
             uiTaskList.addAll(taskList);
             adapter.setAndChangeData(uiTaskList);
+            swipeRefreshLayout.setCanLoadMore(currentIndex == TaskFragment.MY_DONE && (uiTaskList.size() % 12 == 0));
+            noResultText.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
+            taskNoResultImageView.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
         }
 
         @Override
@@ -344,6 +337,7 @@ public class TaskListFragment extends Fragment {
             swipeRefreshLayout.setRefreshing(false);
             WebServiceMiddleUtils.hand(getActivity(), error, errorCode);
             noResultText.setVisibility(taskList.size() > 0 ? View.GONE : View.VISIBLE);
+            taskNoResultImageView.setVisibility(taskList.size() > 0 ? View.GONE : View.VISIBLE);
         }
 
         @Override
@@ -354,14 +348,16 @@ public class TaskListFragment extends Fragment {
                 uiTaskList.remove(deletePosition);
                 adapter.notifyDataSetChanged();
             }
-            noResultText.setVisibility(taskList.size() > 0 ? View.GONE : View.VISIBLE);
+            noResultText.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
+            taskNoResultImageView.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
         }
 
         @Override
         public void returnDeleteTaskFail(String error, int errorCode) {
             swipeRefreshLayout.setRefreshing(false);
             WebServiceMiddleUtils.hand(getActivity(), error, errorCode);
-            noResultText.setVisibility(taskList.size() > 0 ? View.GONE : View.VISIBLE);
+            noResultText.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
+            taskNoResultImageView.setVisibility(uiTaskList.size() > 0 ? View.GONE : View.VISIBLE);
         }
 
     }
